requirements.txt
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

